nexus-app/kernel/requirements.lock.txt
# Nexus Hybrid Kernel — Pinned Dependencies (Generated 2026-02-11)
# Install: pip install -r requirements.lock.txt

aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.2
faster-whisper==1.2.1
filelock==3.20.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
lancedb==0.26.1
langchain==1.2.0
langchain-core==1.2.6
langchain-ollama==1.0.1
numpy==2.4.0
ollama==0.6.1
onnxruntime==1.23.2
openwakeword==0.6.0
orjson==3.11.5
packaging==25.0
piper-tts==1.3.0
psutil==7.2.1
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
sounddevice==0.5.3
soupsieve==2.8.1
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
websockets>=12.0
cryptography>=41.0.0
zeroconf>=0.50.0
